refactor to new path_utils module
